Move `Async::Stop` to `lib/async/stop.rb` and implement compatibility for older Ruby versions.
